WOLF-300: Eliminate (some) undefined behavior from filesystem tests.
libs/filesystem/test/operations_test.cpp had a test which depended on
dereferencing a null (0) pointer. This has been undefined behavior forever;
however it was a "benign" dereferencing, in which (*0) became a reference
which was then passed to the & operator in hopes of recovering the 0 pointer.
That has stopped working with recent versions of clang.

The point was to test with a system::error_category instance with an address
distinct from everything else. Accomplish that alternatively by introducing a
static_category() function that returns a reference to a function-static
instance of a bogus_error_category subclass.

Empirically, this makes the tests pass on clang again.