DRTVWR-418: Consistently use boost_context_preserve_fpu for jump_fcontext().
The boost_context_preserve_fpu mechanism intends to ensure passing the same
bool parameter to every jump_fcontext() call in the process. However, one of
the jump_fcontext() calls did not use that mechanism, which led to troubles in
the Alchemy viewer project, which uses a newer version of Boost than Linden or
Firestorm. Proactively head off that misfortune. (Thanks Drake!)