Fix utf8_codecvt_facet::do_length() definition.
Sometime between Boost 1.57 and 1.64, the utf8_codecvt_facet::do_length()
problem was fixed by providing both const mbstate_t and non-const mbstate_t
overloads, with one redirecting to the other. Only one overrides the
base-class virtual method, of course, and the const-ness of that parameter in
the base class still depends on the platform and library version. But since
Boost has addressed the problem, we can (must) eliminate that use of the
BOOST_CODECVT_DO_LENGTH_CONST macro.