Suppress test of boost::context::callcc()'s ucontext implementation.
On macOS, this test produces a SIGSEGV. It seems moderately likely that this
is due to the same Apple ucontext problem we encountered years ago:
https://lists.apple.com/archives/darwin-dev/2008/Jan/msg00232.html

But the solution to that problem (#define _XOPEN_SOURCE) appears to cause
other compile problems in Boost.Chrono. Attempting to #define that symbol more
locally doesn't work: either the Chrono problem surfaces, or the SIGSEGV
recurs, the latter presumably because we didn't manage to get the #define in
before the first #include of <sys/_structs.h>.

Since we have no intention of engaging BOOST_USE_UCONTEXT or BOOST_USE_WINFIB,
simply suppress the test_call_native test altogether.