SL-793: Disable Boost.Fiber tests that can't work on Mac.
There are actually two different kinds of tests that cannot work on Mac.
There are NUMA-related tests that depend on a topology() function which only
throws an error. In addition, there's an entire suite that repeats most (all?)
of the other Fiber tests with ucontext, which is not merely deprecated on Mac
but no longer compiles at all.

Disabling the latter suite also disables tests in which Boost.Fiber relies on
native Windows Fibers, but as we have no intention of using that
implementation, should be okay.